add file_cleanup

+++ src/file_cleanup.py
@@ -2,8 +2,11 @@ from pathlib import Path
 
 
 def cleanup_files(filepath: Path) -> None:
-    print(type(filepath))
     try:
-        filepath.unlink()
+        if filepath.is_file():
+            filepath.unlink()
+            print(f"Файл {filepath} успешно удалён")
+        else:
+            print(f"Файл {filepath} не существует или это не файл")
     except Exception as e:
-        print(e)
+        print(f"Ошибка при удалении {filepath}: {e}")
